% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SamplePrevRatioMulti}
\alias{SamplePrevRatioMulti}
\title{Cálculo de Tamaño Muestral y Potencia para Estudios de Razón de Prevalencia}
\usage{
SamplePrevRatioMulti(
  alfa,
  potencia = NULL,
  n = NULL,
  p0 = NULL,
  RP = NULL,
  p1 = NULL,
  IC_superior = NULL,
  IC_inferior = NULL,
  EE = NULL,
  n_previo = NULL,
  r = 1,
  metodo = "rho",
  valores_rho = seq(0, 0.9, by = 0.1)
)
}
\arguments{
\item{alfa}{Nivel de significación}

\item{potencia}{Potencia deseada (opcional si se proporciona n)}

\item{n}{Tamaño muestral (opcional si se proporciona potencia)}

\item{p0}{Prevalencia en el grupo no expuesto (requerido si metodo = "rho")}

\item{RP}{Razón de Prevalencia (opcional si se proporciona p1)}

\item{p1}{Prevalencia en el grupo expuesto (opcional si se proporciona RP)}

\item{IC_superior}{Límite superior del intervalo de confianza de la RP}

\item{IC_inferior}{Límite inferior del intervalo de confianza de la RP (opcional si se proporciona EE)}

\item{EE}{Error estándar del coeficiente (opcional si se proporcionan los IC)}

\item{n_previo}{Tamaño de muestra del estudio previo (requerido para método "ee")}

\item{r}{Razón de tamaños de grupo (expuestos/no expuestos), por defecto es 1}

\item{metodo}{Método de cálculo: "rho" o "ee" (error estándar)}

\item{valores_rho}{Vector de valores rho para calcular tamaños muestrales (si metodo = "rho")}
}
\value{
Un data frame con resultados según el método elegido
}
\description{
Cálculo de Tamaño Muestral y Potencia para Estudios de Razón de Prevalencia
}
